Fix DataKey variant limits and harden oracle admin tests for #799.
Nest emergency and governance storage keys to stay within Soroban enum limits, fix benji mint client, and add admin-auth coverage for oracle configuration setters.

+++ contracts/vault/src/permissions.rs
@@ -59,10 +59,10 @@ impl MultiSignerValidator {
         if threshold == 0 {
             return Err("threshold must be > 0");
         }
-        if threshold as usize > signers.len() {
+        if (threshold as u32) > signers.len() {
             return Err("threshold exceeds signer set size");
         }
-        if approvals.len() < threshold as usize {
+        if approvals.len() < threshold {
             return Err("insufficient approvals");
         }
